coinmarketcap_fetcher/requirements.txt
pandas
requests
google-cloud-secret-manager
google-cloud-storage
